requirements.txt
discord.py==2.5.2
httpx==0.28.1
python-dotenv==1.0.1
